create header component
